Add file to parse lisp code
Soon, we'll evaluate the functions recursively, or something. Right now the plan
is to register functions as a string in a dictionary using a decorator, but I
might need some sort of additional information about each function lik
parameters, return value, etc so that type safety can be reasoned about in the
implementation of the genetic program.